+++ src/service/reserve.ts
@@ -1,6 +1,7 @@
 /**
  * 予約サービス
  */
+import * as cinerinoapi from '@cinerino/api-nodejs-client';
 import * as cinerino from '@cinerino/domain';
 import * as moment from 'moment';
 
@@ -217,7 +218,7 @@ export function onReservationStatusChanged(
             // 集計タスク作成
             const aggregateTask: factory.task.aggregateEventReservations.IAttributes = {
                 name: <any>factory.taskName.AggregateEventReservations,
-                project: { typeOf: 'Project', id: params.project.id },
+                project: { typeOf: cinerinoapi.factory.organizationType.Project, id: params.project.id },
                 status: factory.taskStatus.Ready,
                 // Chevreの在庫解放が非同期で実行されるのでやや時間を置く
                 // tslint:disable-next-line:no-magic-numbers