+++ src/service/task/createReturnOrderReport.ts
@@ -12,7 +12,7 @@ import * as AggregateService from '../aggregate';
 export function call(data: factory.task.createReturnOrderReport.IData): IOperation<void> {
     return async (settings: IConnectionSettings) => {
         const aggregateSaleRepo = new AggregateSaleRepo(settings.connection);
-        await AggregateService.report4sales.createReturnOrderReport(<any>data)(
+        await AggregateService.report4sales.createReturnOrderReport(data)(
             aggregateSaleRepo
         );
     };