+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - イベントに対する全注文返品処理において、該当注文の検索を注文取引検索から注文検索へ変更
+- update @cinerino/factory
 
 ### Deprecated